requirement.txt
argparse
bitcoin
coincurve
cryptography
pycryptodome
gevent
pyopenssl
pyyaml
repoze.lru
upnpy
rlp
sympy
base58
ecdsa
mnemonic
sha256